remove Home file, and some print statements
